requirements.txt
# Core ML and RL libraries
torch>=1.12.0
stable-baselines3>=1.6.0
gymnasium>=0.26.0
numpy>=1.21.0
pandas>=1.4.0
scikit-learn>=1.1.0

# Data processing and technical analysis
ta>=0.10.0
yfinance>=0.1.87
ccxt>=2.0.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=0.19.0
requests>=2.28.0
aiohttp>=3.8.0

# Logging and monitoring
structlog>=22.1.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-mock>=3.8.0

# Development tools
black>=22.0.0
flake8>=5.0.0
mypy>=0.991